tests/Node-cloneNode.html

<!DOCTYPE html>
<title>Node.cloneNode</title>
<script src="testharness.js"></script>
<div id=log></div>
<script>
test(function() {
  var el = document.createElement("foo");
  el.setAttribute("a", "b");
  el.setAttribute("c", "d");
  var c = el.cloneNode(false);
  assert_true(c.attributes != el.attributes);
  assert_true(c.attributes !== el.attributes);
  assert_true(c.attributes[0] !== el.attributes[0]);
  assert_equals(c.attributes[0].name, el.attributes[0].name);
  assert_equals(c.attributes[0].prefix, el.attributes[0].prefix);
  assert_equals(c.attributes[0].namespaceURI, el.attributes[0].namespaceURI);
  assert_equals(c.attributes[0].value, el.attributes[0].value);
})
</script>
